Frontend major impact - разделил authmanager на два варианта: серверный и клиентский -

+++ backend/src/main/java/com/atlas/service/PostService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 
+import com.atlas.constant.Role;
 import com.atlas.model.Post;
 import com.atlas.repository.PostRepository;
 
@@ -13,6 +14,8 @@ import com.atlas.repository.PostRepository;
 public class PostService {
     @Autowired
     private PostRepository postRepository;
+    @Autowired
+    private JwtService jwtService;
     private final int perpage = 12;
 
     public Post findById(int id) {
@@ -28,9 +31,31 @@ public class PostService {
         return postRepository.save(post);
     }
 
-    public boolean delete(int id) {
+    public boolean delete(int id, String jwt) {
+        String userrole = jwtService.getUserRole(jwt);
+        if (userrole == null || !userrole.equals(Role.ADMIN)) {
+            return false;
+        }
+
         if (!postRepository.existsById(id)) return false;
         postRepository.deleteById(id);
+        return true;
+    }
+
+    public boolean edit(int id, String jwt, String body) {
+        String userrole = jwtService.getUserRole(jwt);
+        if (userrole == null || !(userrole.equals(Role.ADMIN) || userrole.equals(Role.MANAGER))) {
+            return false;
+        }
+        if (!postRepository.existsById(id)) {
+            return false;
+        }
+        Post postFromDB = postRepository.findById(id).get();
+        if (postFromDB == null) {
+            return false;
+        }
+        postFromDB.setBody(body);
+        postRepository.save(postFromDB);
         return true;
     }
